refactor(createApiModel): reusable function for api model creation
