+++ packages/website/src/routes/docs/$branchName/packages/$packageName/index.tsx
@@ -1,27 +1,15 @@
-/* eslint-disable @typescript-eslint/no-unsafe-member-access */
-/* eslint-disable @typescript-eslint/no-unsafe-assignment */
 import { json } from '@remix-run/node';
 import { Params, useLoaderData } from '@remix-run/react';
-import { ApiItem, ApiModel, ApiPackage } from '~/api-extractor.server';
-import { TSDocConfiguration } from '~/tsdoc.server';
+import { createApiModel } from '~/util/api-model.server';
 import { findPackage, getMembers } from '~/util/parse.server';
 
 export async function loader({ params }: { params: Params }) {
 	const res = await fetch(
 		`https://raw.githubusercontent.com/discordjs/docs/main/${params.packageName!}/${params.branchName!}.api.json`,
 	);
+	// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
 	const data = await res.json();
-
-	const model = new ApiModel();
-	// eslint-disable-next-line @typescript-eslint/no-unsafe-argument
-	const apiPackage = ApiItem.deserialize(data, {
-		apiJsonFilename: '',
-		toolPackage: data.metadata.toolPackage,
-		toolVersion: data.metadata.toolVersion,
-		versionToDeserialize: data.metadata.schemaVersion,
-		tsdocConfiguration: new TSDocConfiguration(),
-	}) as ApiPackage;
-	model.addMember(apiPackage);
+	const model = createApiModel(data);
 
 	const pkg = findPackage(model, params.packageName!);
 	return json({
