+++ packages/website/src/tsdoc-config.server.ts
@@ -1,0 +1,1 @@
+export * from '@microsoft/tsdoc-config';
